Add stream RPC client
