+++ x10/clients/streamrpc/streamrpc_client.py
@@ -1,24 +1,23 @@
 import asyncio
 import json
 import random
-from typing import Any, Callable, Coroutine, TypeAlias, TypeVar
+from typing import Any, Callable, Coroutine, TypeVar
 
 import websockets
+from clients.streamrpc.streamrpc_dispatcher import (
+    OnSequenceBreakCallback,
+    PendingRequestsMap,
+)
 from websockets import ConnectionClosed
 
+from x10.clients.streamrpc.streamrpc_dispatcher import RequestId, StreamRpcDispatcher
 from x10.clients.streamrpc.subscription_params import (
     StreamMessageHandler,
     SubscribeParams,
     TopicId,
     TopicSubscription,
 )
-from x10.errors import (
-    StreamRpcConnectionError,
-    StreamRpcError,
-    StreamRpcServerError,
-    StreamRpcTimeoutError,
-)
-from x10.models.stream_rpc import StreamRpcResponseModel
+from x10.errors import StreamRpcConnectionError, StreamRpcError, StreamRpcTimeoutError
 from x10.utils.http import USER_AGENT, RequestHeader
 from x10.utils.log import get_logger
 
@@ -26,9 +25,7 @@ LOGGER = get_logger(__name__)
 CONNECTION_LOOP_TASK_NAME = "x10-rpc-connection-loop"
 
 T = TypeVar("T")
-RequestId: TypeAlias = str
 OnReconnectCallback = Callable[[list[str]], Coroutine[Any, Any, None]]
-OnSequenceBreakCallback = Callable[[str, int, int], Coroutine[Any, Any, None]]
 
 
 class StreamRpcClient:
@@ -111,8 +108,7 @@ class StreamRpcClient:
         """
 
         result = await self._rpc("list-subscriptions")
-        # FIXME: Simplify?
-        return result.get("subscriptions") or []
+        return result["subscriptions"]
 
     async def subscribe(self, *, params: SubscribeParams[T], handler: StreamMessageHandler):
         """
@@ -175,10 +171,6 @@ class StreamRpcClient:
         # FIXME: Replace with state?
         self._is_stopped = False
         self._next_request_id = 0
-        # FIXME: Update description
-        # Last observed connection-level sequence number; None until the first
-        # message arrives on a connection (also reset to None on each reconnect).
-        self._last_seq: int | None = None
 
         # FIXME: Update description
         # Fires when a connection is fully established (and resubscription done).
@@ -189,11 +181,17 @@ class StreamRpcClient:
 
         # FIXME: Update description
         # Pending RPC request futures keyed by request id.
-        self._pending_requests: dict[RequestId, asyncio.Future[dict[str, Any]]] = {}
+        self._pending_requests: PendingRequestsMap = {}
 
         # FIXME: Update description
         # Active subscriptions keyed by topic_id.
         self._subscriptions: dict[TopicId, TopicSubscription] = {}
+
+        self._dispatcher = StreamRpcDispatcher(
+            pending_requests=self._pending_requests,
+            subscriptions=self._subscriptions,
+            on_sequence_break=on_sequence_break,
+        )
 
     async def __aenter__(self) -> "StreamRpcClient":
         await self.connect()
@@ -295,7 +293,7 @@ class StreamRpcClient:
                     LOGGER.debug("Connected to %s", self._api_url)
 
                     # `seq` restarts at 0 on each new connection
-                    self._last_seq = None
+                    self._dispatcher.reset_last_seq()
                     reconnect_delay = self._reconnect_initial_delay
 
                     await self._resubscribe()
@@ -308,7 +306,7 @@ class StreamRpcClient:
 
                     async for raw in ws:
                         if isinstance(raw, str):
-                            self._dispatch_raw(raw)
+                            self._dispatcher.dispatch_raw(raw)
             except asyncio.CancelledError:
                 break
             except (ConnectionClosed, OSError, asyncio.TimeoutError) as exc:
@@ -356,111 +354,3 @@ class StreamRpcClient:
                 await self._ws.send(json.dumps(request))
             except Exception:
                 LOGGER.exception("Failed to resubscribe to %s", topic_id)
-
-    # FIXME: Create a dispatcher class?
-    def _dispatch_raw(self, raw: str) -> None:
-        """
-        Parse a raw WebSocket text frame and route it to the right handler.
-        """
-
-        try:
-            msg: dict[str, Any] = json.loads(raw)
-        except json.JSONDecodeError:
-            LOGGER.warning("Received invalid JSON (%.120s…)", raw)
-            return
-
-        # (1) JSON-RPC response
-        request_id: RequestId | None = msg.get("id")
-
-        if request_id is not None:
-            # FIXME: Create a class instance?
-            request_result = self._pending_requests.get(str(request_id))
-
-            if not request_result:
-                LOGGER.warning("Received response for unknown request id=%s", request_id)
-                return
-
-            err = msg.get("error")
-
-            if err:
-                request_result.set_exception(
-                    StreamRpcServerError(code=err["code"], message=err["message"], data=err.get("data"))
-                )
-            else:
-                request_result.set_result(msg["result"])
-
-            return
-
-        # (2) Stream data
-        # FIXME: Create a class instance?
-        subscription_id: str | None = msg.get("subscription")
-
-        if subscription_id is not None:
-            asyncio.ensure_future(self._dispatch_message(msg, subscription_id))
-            return
-
-        # (3) Unknown message
-        LOGGER.error("Unrecognised message shape: %s", raw)
-
-    async def _dispatch_message(self, msg: dict[str, Any], subscription_id: str) -> None:
-        """
-        Deserialize a stream message and invoke the registered handler.
-        """
-
-        subscription = self._subscriptions.get(subscription_id)
-
-        if subscription is None:
-            LOGGER.warning("Received message for unknown subscription id=%s", subscription_id)
-            return
-
-        msg_seq = msg["seq"]
-
-        if self._last_seq is not None and msg_seq != self._last_seq + 1:
-            LOGGER.warning(
-                "Sequence break detected for subscription %s: last_seq=%s, msg_seq=%s",
-                subscription_id,
-                self._last_seq,
-                msg_seq,
-            )
-
-            if self._on_sequence_break:
-                try:
-                    result = await self._on_sequence_break(subscription_id, self._last_seq, msg_seq)
-
-                    if asyncio.iscoroutine(result):
-                        await result
-                except Exception:
-                    LOGGER.exception("Unhandled exception in `on_sequence_break` callback")
-
-        self._last_seq = msg_seq
-
-        msg_data = msg["data"]
-        msg_type = msg["type"]
-
-        try:
-            deserialized_data = subscription.params.deserialize_data(msg_data, msg_type)
-        except Exception as exc:
-            LOGGER.exception(
-                "Failed to deserialize message for subscription %s (type=%s, seq=%s): %s",
-                subscription_id,
-                msg_type,
-                msg_seq,
-                exc,
-            )
-            return
-
-        enveloped_data = StreamRpcResponseModel(
-            type=msg_type,
-            data=deserialized_data,
-            ts=msg["ts"],
-            seq=msg_seq,
-            subscription=subscription_id,
-        )
-
-        try:
-            result = subscription.handler(enveloped_data)
-
-            if asyncio.iscoroutine(result):
-                await result
-        except Exception:
-            LOGGER.exception("Unhandled exception in handler for subscription %s", subscription_id)
